// src/app/mario/interfaces/serialized-level.ts
export interface SerializedLevel {

    key: string
    id: number
    line_00?: string[]
    line_01?: string[]
    line_02?: string[]
    line_03?: string[]
    line_04?: string[]
    line_05?: string[]
    line_06?: string[]
    line_07?: string[]
    line_08?: string[]
    line_09?: string[]
    line_10?: string[]
    line_11?: string[]
    line_12?: string[]
    line_13?: string[]
    line_14?: string[]

}